Refactored pendulum examples
