+++ CPlusPlus/Templates/Makefile
@@ -9,6 +9,15 @@ APPL = swap.exe sum.exe points.exe pendulum_func_obj.exe pendulum_bind.exe \
 
 all: $(APPL)
 
+pendulum_func_obj.exe: pendulum_func_obj.o pendulum_utils.o
+	$(CXX) $(CXXFLAGS) -o $@ $^ $(LIBS)
+
+pendulum_bind.exe: pendulum_bind.o pendulum_utils.o
+	$(CXX) $(CXXFLAGS) -o $@ $^ $(LIBS)
+
+pendulum_lambda.exe: pendulum_lambda.o pendulum_utils.o
+	$(CXX) $(CXXFLAGS) -o $@ $^ $(LIBS)
+
 %.exe: %.o
 	$(CXX) $(CXXFLAGS) -o $@ $^ $(LIBS)
