+++ CPlusPlus/Templates/pendulum_lambda.cpp
@@ -1,7 +1,7 @@
 #include <cmath>
 #include <functional>
-#include <iostream>
-#include <unistd.h>
+
+#include "pendulum_utils.h"
 
 const double pi {acos(-1.0)};
 
@@ -9,48 +9,10 @@ double pendulum_func(double t, double freq) {
     return cos(2.0*pi*freq*t);
 }
 
-void print_help();
-
 int main(int argc, char *argv[]) {
-    double freq {1.0};
-    double max_t {1.0};
-    double delta_t {0.005};
-    char opt {'\0'};
-    while ((opt = getopt(argc, argv, "f:d:m:h")) != -1) {
-        switch (opt) {
-            case 'f':
-                freq = std::stof(optarg);
-                break;
-            case 'd':
-                delta_t = std::stof(optarg);
-                break;
-            case 'm':
-                max_t = std::stof(optarg);
-                break;
-            case 'h':
-                print_help();
-                return 0;
-            default:
-                std::cerr << "# error: unknown option" << std::endl;
-                print_help();
-                return 1;
-        }
-    }
-    using namespace std::placeholders;
+    double freq, delta_t, max_t;
+    std::tie(freq, delta_t, max_t) = get_parameters(argc, argv);
     auto pendulum = [freq](double t) { return pendulum_func(t, freq); };
-    for (double t = 0.0; t <= max_t; t += delta_t)
-        std::cout << t << "," << pendulum(t) << std::endl;
+    integrate(pendulum, delta_t, max_t);
     return 0;
-}
-
-void print_help() {
-    std::cerr << "Usage: pendulum_bind.exe "
-              << "[ -f <freq> ] "
-              << "[ -d <delta_t> ] "
-              << "[ -m <max_t> ] "
-              << std::endl;
-    std::cerr << "\t-f <freq>    : pendulum frequency" << std::endl;
-    std::cerr << "\t-d <delta_t> : time increment" << std::endl;
-    std::cerr << "\t-m <max_t>   : maximum time for the pendulum"
-              << std::endl;
 }
